[fix] CardHistoryコンポーネントを修正
