+++ front/app/components/CardHistory/CardHistory.module.css
@@ -1,0 +1,30 @@
+.card {
+  background-color: #3d6f5d; /* #65B741 */
+  color: rgb(251, 246, 238); /* #FBF6EE */
+  padding: 1.5rem;
+  width: auto;
+  border-radius: 0.5rem;
+}
+
+.card-history {
+  background-color: #3d6f5d; /* #65B741 */
+  color: rgb(251, 246, 238); /* #FBF6EE */
+  padding: 1rem;
+  width: auto;
+  border-radius: 0.5rem;
+}
+
+.card-title {
+  font-family: 'BIZ UDGothic', 'sans-serif';
+  font-weight: bold;
+  font-size: 2.5rem;
+  text-decoration: underline;
+  text-underline-offset: 0.8rem;
+}
+
+.card-content {
+  margin-top: 1.5rem;
+  font-family: 'BIZ UDGothic', 'sans-serif';
+  font-size: 2rem;
+  min-height: 40vh; /* content 部分の最小の高さを設定 */
+}
