+++ front/app/components/CardHistory/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CardHistory';
